chore: merge generate-binary into generate
