+++ packages/cli-tools/bin/streamr-mock-data.ts
@@ -6,6 +6,5 @@ program
     .version(pkg.version)
     .usage('<command> [<args>]')
     .description('mock-data subcommands')
-    .command('generate', 'generate JSON data')
-    .command('generate-binary', 'generate binary data')
+    .command('generate', 'generate random JSON or binary data')
     .parse()
